fix(D08-1): 修复 12 处编译错误 + 1 处警告
- E0507 inventory_move.rs: 移除 txn.rollback()，事务 Drop 时自动回滚
- E0308 po/receipt.rs: 接住 create_stock_fabric_txn 返回的 Model
- E0308 purchase_receipt_private.rs: items 改为 &mut order_item_map 借用
- E0308 quotation_service.rs: build_item/term_active_models 的 quotation_id 改为 i64
- E0382 material_shortage_service.rs: 删除未使用的 _threshold 死代码
- E0592 ar/vfy_ops/match.rs: 重命名为 group_invoice_refs_by_customer 避免与 aging.rs 重复
- E0308 dashboard_service.rs: cache_key 转 String 后再借用
- E0308 report/tpl.rs: make_col width 参数改为 Option<u32>
- warning: ar_ops/verification_ops/manual.rs 移除未使用的 now 参数

+++ backend/src/services/quotation_service.rs
@@ -150,7 +150,7 @@ impl QuotationService {
     /// 构建报价单明细 ActiveModel 列表
     fn build_item_active_models(
         items: &[CreateQuotationItemDto],
-        quotation_id: i32,
+        quotation_id: i64,
     ) -> Vec<ItemActive> {
         let mut models: Vec<ItemActive> = Vec::with_capacity(items.len());
         for (idx, item_dto) in items.iter().enumerate() {
@@ -201,7 +201,7 @@ impl QuotationService {
     /// 构建报价单贸易条款 ActiveModel 列表
     fn build_term_active_models(
         terms: Vec<CreateQuotationTermDto>,
-        quotation_id: i32,
+        quotation_id: i64,
     ) -> Vec<TermActive> {
         terms
             .into_iter()
